위치 보정 정밀도 강화 — 캐릭터가 선 자리에 더 정확히 수렴
- 미니맵 점을 서브픽셀 centroid(소수점)로 측정 — 1px 미만 이동도 감지,
  코스 허용오차 1px → 0.6px (미니맵 1px ≈ 실좌표 수 px이므로 체감 큼)
- 파인 허용오차 4px → 2px, 잔여 오차 6px 이하에선 비례 홀드 대신
  28ms 최소 탭으로 콕콕 이동(오버슈트 방지)
- 코스/파인 모두 목표를 지나쳐 부호가 뒤집히면 다음 탭 세기를 55%로
  감쇠 — 좌우 진동 없이 빠르게 수렴
- 탭 후 재측정 대기 150ms → 220ms: 키를 뗀 뒤 캐릭터가 미끄러져 완전히
  멈춘 상태에서 재기 때문에 측정 자체가 정확해짐
- 기존 watcher.json의 구 기본값(1/4/150)은 로드 시 새 기본값으로 마이그레이션

+++ src/YInput.Host/Web/ApiEndpoints.cs
@@ -453,7 +453,7 @@ public static class ApiEndpoints
     private sealed record OverlayTargetBody(string? Process = null);
     private sealed record WatcherBody(string? Process = null, int? TolerancePx = null,
                                       double? MsPerPx = null, int? MaxCorrectionMs = null, double? MinScore = null,
-                                      int? MiniTolerancePx = null, double? MsPerMiniPx = null);
+                                      double? MiniTolerancePx = null, double? MsPerMiniPx = null);
     private sealed record RegionBody(int X = 0, int Y = 0, int W = 0, int H = 0);
     private sealed record RecordStartBody(
         bool Keyboard = true,
